Extract escapeHtml to shared utility in utils/html-sanitizer.js
Moves the escapeHtml function from auth/oauth-server.js to the shared
html-sanitizer module so it can be reused across the codebase.

Refs ryaker/outlook-mcp#42
